removed unused tests code.
