+++ tests/unit/backend/layers/common/base_api_test.py
@@ -4,8 +4,8 @@ import time
 import unittest
 from unittest.mock import patch
 from backend.layers.thirdparty.cdn_provider_interface import CDNProviderInterface
-from tests.unit.backend.api_server.config import TOKEN_EXPIRES
 from tests.unit.backend.layers.common.base_test import BaseTest
+from tests.unit.backend.layers.api.config import TOKEN_EXPIRES
 
 
 class BaseAuthAPITest(unittest.TestCase):
