+++ tests/unit/backend/layers/common/test_auth0_manager.py
